fpl 1.4: embed local headers
Fpl is no longer a subproject of the jest project.  Therefore,
the generated code can't assume it has jest/fpl headers available.

So, instead of generating code which #includes local project headers,
@include the headers, which embeds them in the generated code.

There's one problem with that, and that is headers can include headers,
and jemp doesn't know about that.  So, the hack fix is to #ifndef
around local included sub-headers and brute-force @include all dependencies
of any other @included headers.  It's not great but it works.

Standard library headers are still #included as usual.