Fix issue 4; change SRC_DIR, DATA_DIR envs
Fix to issue #4 wherin there was a full path in a test expectation:
  https://github.com/ccomparini/fpl2cc/issues/4

Ooops.

Changed cc test env variables SRC_DIR and DATA_DIR to both be
relative paths, and added BASE_DIR as an absolute path and the
base for the SRC_DIR and DATA_DIRs for cases where an absolute
path may be necessary or helpful.