Remove an XXX in the repetition test and move on
The XXX correctly points out a flaw in how it works if you write
custom separator code in c++.  However, that has nothing to do
with the test itself, and anyway I'm not sure one should be writing
separator code in c++ anyway.  I'd like to phase out embedded c++
entirely, and separators would be a good place to start, now that
there are other ways of writing them.