"fix" line numbers for missing return warnings
The jemp generally inserts #line preprocessor directives in the
generated code.  This is normally correct, but in the case where
the fpl author has neglected to include a return statement in
their reduce handler, it causes the c++ compiler to report the
line in the jemp instead of in the line in the fpl.  So jam in
a special #line directive at the end of the reduce action jemp
template to point it at the actual source fpl.