Fix issue 14.  Much custom terminal rework
Rework custom terminals such that we don't integrate all the .jemp
into productions.h.  This introduces a new struct to hold the
data to use when filling in the jemp for custom terminals.

Add an explicit grammar element type for assertions.

Finally (the point of this) - fix issue 14 - @terminal now uses
the same method to parse basic terminals as @scanner so @terminals
can be specified using regexes etc.