Add .gitignore; generate .gitignore in build dir
Add a .gitignore for a couple of cache files which show up in
"git status".

Also make the top level SConstruct add a .gitignore file to the
build dir.  This could be hard-coded in the top level .gitignore
(which would actually be simpler) but I decided to put it here
so that it always gets put in the actual build directory, even
if the build dir gets renamed.  Moving on...