subprof (yprof):  check if we're really profiling the current HEAD
Check if the subprof is running against the current HEAD, and, if not,
let the user know and put an asterisk after the commit ID in the result
file.

I figure that, in general, we won't want to check in yprof results
which don't have corresponding commit IDs, but the HEAD should still
be the closest commit ID, and in many cases will match the code
which was run, so adding an asterisk to the ID seems like a good way
to warn that the commit ID might not line up exactly in cases where
the yprof gets checked in.