Support !! and !? error/warnings on rule matches.
Add the ability to make rules issue errors or warnings when matched.

The classic cases where one might use this would be reporting
unterminated strings or mismatched braces (errors), or deprecated/
potentially dangerous language features and usage (warnings).

The broader idea is to add the ability to issue better error
messages and warnings in generated languages, particularly at
the abstract fpl level so that specific implementations don't
have to worry about details.

Incidentally, this separates the templates for reduce actions
such that they're organized by implementation language (though
only DEFAULT actually exists as yet), because that's how I though
I was going to implement error messages.  This may or may not be
useful, but it's not worse than before (and possibly better), so
I'm keeping it for now.